refactor commands for better separation
