+++ src/command/client/event.rs
